fix: Use separate templates for pages and facts.
This implements a PageWrapper component for the reusable styles, but 
uses slightly different styling depending on the type of content being 
shown. Also fixes the issue where pages were showing an error.
